refactor: VatWorkerClient -> VatWorkerManager
