+++ packages/nodejs/src/index.ts
@@ -1,2 +1,2 @@
-export { NodejsVatWorkerClient } from './kernel/VatWorkerClient.ts';
+export { NodejsVatWorkerManager } from './kernel/VatWorkerManager.ts';
 export { makeKernel } from './kernel/make-kernel.ts';
