+++ packages/nodejs/test/e2e/VatWorkerManager.test.ts
@@ -1,57 +1,42 @@
 import '@ocap/shims/endoify';
 
 import type { VatId } from '@ocap/kernel';
+import { NodeWorkerDuplexStream } from '@ocap/streams';
 import { makeCounter } from '@ocap/utils';
 import { describe, expect, it, vi } from 'vitest';
 
-import { NodejsVatWorkerClient } from './VatWorkerClient.ts';
+import { NodejsVatWorkerManager } from '../../src/kernel/VatWorkerManager.ts';
+import { getTestWorkerFile } from '../get-test-worker.ts';
 
-const mocks = vi.hoisted(() => ({
-  worker: {
-    once: (_: string, callback: () => unknown) => {
-      callback();
-    },
-    terminate: vi.fn(async () => undefined),
-  },
-  stream: {
-    synchronize: vi.fn(async () => undefined).mockResolvedValue(undefined),
-    return: vi.fn(async () => ({})),
-  },
-}));
-
-vi.mock('@ocap/streams', () => ({
-  NodeWorkerDuplexStream: vi.fn(() => mocks.stream),
-}));
-
-vi.mock('node:worker_threads', () => ({
-  Worker: vi.fn(() => mocks.worker),
-}));
-
-describe('NodejsVatWorkerClient', () => {
-  it('constructs an instance without any arguments', () => {
-    const instance = new NodejsVatWorkerClient({});
-    expect(instance).toBeInstanceOf(NodejsVatWorkerClient);
-  });
-
-  const workerFilePath = 'unused';
+describe('NodejsVatWorkerManager', () => {
+  const testWorkerFile = getTestWorkerFile('stream-sync');
   const vatIdCounter = makeCounter();
   const getTestVatId = (): VatId => `v${vatIdCounter()}`;
 
   describe('launch', () => {
     it('creates a NodeWorker and returns a NodeWorkerDuplexStream', async () => {
-      const service = new NodejsVatWorkerClient({
-        workerFilePath,
+      const service = new NodejsVatWorkerManager({
+        workerFilePath: testWorkerFile,
       });
       const testVatId: VatId = getTestVatId();
       const stream = await service.launch(testVatId);
 
-      expect(stream).toStrictEqual(mocks.stream);
+      expect(stream).toBeInstanceOf(NodeWorkerDuplexStream);
     });
 
     it('rejects if synchronize fails', async () => {
       const rejected = 'test-reject-value';
-      mocks.stream.synchronize.mockRejectedValue(rejected);
-      const service = new NodejsVatWorkerClient({ workerFilePath });
+
+      vi.doMock('@ocap/streams', () => ({
+        NodeWorkerDuplexStream: vi.fn().mockImplementation(() => ({
+          synchronize: vi.fn(() => 'no').mockRejectedValue(rejected),
+        })),
+      }));
+      vi.resetModules();
+      const NVWS = (await import('../../src/kernel/VatWorkerManager.ts'))
+        .NodejsVatWorkerManager;
+
+      const service = new NVWS({ workerFilePath: testWorkerFile });
       const testVatId: VatId = getTestVatId();
       await expect(async () => await service.launch(testVatId)).rejects.toThrow(
         rejected,
@@ -61,8 +46,8 @@ describe('NodejsVatWorkerClient', () => {
 
   describe('terminate', () => {
     it('terminates the target vat', async () => {
-      const service = new NodejsVatWorkerClient({
-        workerFilePath,
+      const service = new NodejsVatWorkerManager({
+        workerFilePath: testWorkerFile,
       });
       const testVatId: VatId = getTestVatId();
 
@@ -74,8 +59,8 @@ describe('NodejsVatWorkerClient', () => {
     });
 
     it('throws when terminating an unknown vat', async () => {
-      const service = new NodejsVatWorkerClient({
-        workerFilePath,
+      const service = new NodejsVatWorkerManager({
+        workerFilePath: testWorkerFile,
       });
       const testVatId: VatId = getTestVatId();
 
@@ -87,8 +72,8 @@ describe('NodejsVatWorkerClient', () => {
 
   describe('terminateAll', () => {
     it('terminates all vats', async () => {
-      const service = new NodejsVatWorkerClient({
-        workerFilePath,
+      const service = new NodejsVatWorkerManager({
+        workerFilePath: testWorkerFile,
       });
       const vatIds: VatId[] = [getTestVatId(), getTestVatId(), getTestVatId()];
